Add Raspberry Pi EEPROM updater and config patching
This commit adds logic to handle Raspberry Pi firmware in pi-topOS.

On install, `rpi-eeprom-update` systemd service is disabled in favour of `pt-rpi-eeprom-update`. `/etc/default/rpi-eeprom-update` is backed up and then modified to point to `/recovery` and add a hook that takes in the current bootloader config as input on stdin, and prints out a modified config to stdout, depending on the device. This hook is used by `rpi-eeprom-update`, as well as a wrapper that was written to check that the config is correct, even when there is no update available.

This ensures that a pi-top [4] running pi-topOS will always run with the correct bootloader config.

If the user is running on a Raspberry Pi, then EEPROM config will not be changed.

If a config change is required for the current session, then a reboot will occur. In all other circumstances, a notification will be shown to the user that an update is ready to be applied with a reboot.

Clean up of old files in `/recovery` is handled explicitly, and VL805 updates are done separately to the bootloader.

NOTE: the systemd service is only enabled on install. It does not start on install or restart on upgrade in order to prevent the potential situation of rebooting in the middle of an update.

The systemd service runs on multi-user.target in order to run at the same time as the `rpi-eeprom-update` service would, but needs to wait to be able to determine the user that is using the display. Therefore, it waits for this information before attempting to show a notification. It will time out after 10 seconds.