Add support for new alsa configuration of RPi audio devices
`pt-os-mods` now includes a generic systemd service which can be appropriated for any use case that might require multiple runs or are to be run contextually (as opposed to running at the moment of package installation via `postinst`).

New kernel (>5) has changed how audio appears in Alsa. This commit aims to ensure that audio experience for users is seamless (within reason).

Further changes are likely required for peripherals to continue to work - this will be handled in the device manager and common library as appropriate. 

We have also added tests for `postinst` and for the new systemd service. 🎉